// database/src/migrations/common/template.enum.ts
export enum COMMON_SURVEY_METHODOLOGY {
  STRATIFIED_RANDOM_BLOCK = 'Stratified Random Block',
  COMPOSITION = 'Composition',
  RECRUITMENT = 'Recruitment',
  TOTAL_COUNT = 'Total Count',
  ENCOUNTER_TRANSECTS = 'Encounter Transects',
  FIXED_WIDTH_TRANSECTS = 'Fixed-width Transects',
  WILDLIFE_CAMERA = 'Wildlife Camera',
  TRACK_COUNT = 'Track Count',
  SPOTLIGHT_COUNT = 'Spotlight Count',
  CLASSIFICATION_TRANSECTS_BLOCKS = 'Classification Transects/Blocks',
  PELLET_SCAT_COUNT = 'Pellet/Scat Count',
  CALL_PLAYBACK = 'Call Playback',
  DNA_INDIVIDUAL = 'DNA - Individual',
  DNA_ENVIRONMENTAL = 'DNA - Environmental',
  MARK_RESIGHT_RECAPTURE = 'Mark Resight Recapture',
  MARK_RESIGHT_RECAPTURE_WILDLIFE_CAMERA = 'Mark Resight Recapture - Wildlife Camera',
  MARK_RESIGHT_RECAPTURE_SPOTLIGHT_COUNT = 'Mark Resight Recapture - Spotlight Count',
  MARK_RESIGHT_RECAPTURE_DNA_INDIVIDUAL = 'Mark Resight Recapture - DNA - Individual',
  DESCRIBED_IN_COMMENTS = 'Described in Comments'
}

export enum TEMPLATE_NAME {
  MOOSE_SRB_OR_COMPOSITION_SURVEY = 'Moose SRB or Composition Survey',
  MOOSE_RECRUITMENT_SURVEY = 'Moose Recruitment Survey',
  MOOSE_COMPOSITION_SURVEY = 'Moose Composition Survey',
  MOOSE_SRB_SURVEY = 'Moose SRB Survey',
  MOOSE_SUMMARY_RESULTS = 'Moose Summary Results',
  SHEEP_COMPOSITION_OR_RECRUITMENT_SURVEY = 'Sheep Composition or Recruitment Survey',
  GOAT_COMPOSITION_OR_RECRUITMENT_SURVEY = 'Goat Composition or Recruitment Survey'
}

export enum SPECIES_NAME {
  MOOSE = 'Moose'
}